+++ packages-content-model/roosterjs-content-model-editor/lib/editor/coreApi/createContentModel.ts
@@ -19,7 +19,7 @@ import {
  * @param selectionOverride When passed, use this selection range instead of current selection in editor
  */
 export const createContentModel: CreateContentModel = (core, option, selectionOverride) => {
-    let cachedModel = selectionOverride ? null : core.contentModelEdit.cachedModel;
+    let cachedModel = selectionOverride ? null : core.cache.cachedModel;
 
     if (cachedModel && core.lifecycle.shadowEditFragment) {
         // When in shadow edit, use a cloned model so we won't pollute the cached one
@@ -32,7 +32,7 @@ export const createContentModel: CreateContentModel = (core, option, selectionOv
         const model = internalCreateContentModel(core, option, selectionOverride);
 
         if (!option && !selectionOverride) {
-            core.contentModelEdit.cachedModel = model;
+            core.cache.cachedModel = model;
         }
 
         return model;
