+++ packages-content-model/roosterjs-content-model-editor/lib/editor/coreApi/getSelectionRangeEx.ts
@@ -7,8 +7,5 @@ import { GetSelectionRangeEx } from 'roosterjs-editor-types';
 export const getSelectionRangeEx: GetSelectionRangeEx = core => {
     const contentModelCore = core as ContentModelEditorCore;
 
-    return (
-        contentModelCore.contentModelEdit.cachedRangeEx ??
-        core.originalApi.getSelectionRangeEx(core)
-    );
+    return contentModelCore.cache.cachedRangeEx ?? core.originalApi.getSelectionRangeEx(core);
 };
