+++ packages-content-model/roosterjs-content-model-editor/lib/editor/coreApi/switchShadowEdit.ts
@@ -14,9 +14,7 @@ export const switchShadowEdit: SwitchShadowEdit = (editorCore, isOn): void => {
 
     if (isOn != !!core.lifecycle.shadowEditFragment) {
         if (isOn) {
-            const model = !core.contentModelEdit.cachedModel
-                ? core.api.createContentModel(core)
-                : null;
+            const model = !core.cache.cachedModel ? core.api.createContentModel(core) : null;
             const range = core.api.getSelectionRange(core, true /*tryGetFromCache*/);
 
             // Fake object, not used in Content Model Editor, just to satisfy original editor code
@@ -36,8 +34,8 @@ export const switchShadowEdit: SwitchShadowEdit = (editorCore, isOn): void => {
 
             // This need to be done after EnteredShadowEdit event is triggered since EnteredShadowEdit event will cause a SelectionChanged event
             // if current selection is table selection or image selection
-            if (!core.contentModelEdit.cachedModel && model) {
-                core.contentModelEdit.cachedModel = model;
+            if (!core.cache.cachedModel && model) {
+                core.cache.cachedModel = model;
             }
 
             core.lifecycle.shadowEditSelectionPath = selectionPath;
@@ -54,8 +52,8 @@ export const switchShadowEdit: SwitchShadowEdit = (editorCore, isOn): void => {
                 false /*broadcast*/
             );
 
-            if (core.contentModelEdit.cachedModel) {
-                core.api.setContentModel(core, core.contentModelEdit.cachedModel);
+            if (core.cache.cachedModel) {
+                core.api.setContentModel(core, core.cache.cachedModel);
             }
         }
     }
