+++ packages-content-model/roosterjs-content-model-editor/lib/editor/plugins/ContentModelEditPlugin.ts
@@ -1,20 +1,13 @@
-import applyDefaultFormat from '../../publicApi/format/applyDefaultFormat';
 import keyboardDelete from '../../publicApi/editing/keyboardDelete';
-import { ContentModelEditPluginState } from '../../publicTypes/pluginState/ContentModelEditPluginState';
 import { IContentModelEditor } from '../../publicTypes/IContentModelEditor';
-import { isCharacterValue } from 'roosterjs-editor-dom';
 import {
+    EditorPlugin,
     IEditor,
     Keys,
     PluginEvent,
     PluginEventType,
     PluginKeyDownEvent,
-    PluginWithState,
-    SelectionRangeEx,
 } from 'roosterjs-editor-types';
-
-// During IME input, KeyDown event will have "Process" as key
-const ProcessKey = 'Process';
 
 /**
  * ContentModel plugins helps editor to do editing operation on top of content model.
@@ -22,17 +15,8 @@ const ProcessKey = 'Process';
  * 1. Delete Key
  * 2. Backspace Key
  */
-export default class ContentModelEditPlugin
-    implements PluginWithState<ContentModelEditPluginState> {
+export default class ContentModelEditPlugin implements EditorPlugin {
     private editor: IContentModelEditor | null = null;
-
-    /**
-     * Construct a new instance of ContentModelEditPlugin class
-     * @param state State of this plugin
-     */
-    constructor(private state: ContentModelEditPluginState) {
-        // TODO: Remove tempState parameter once we have standalone Content Model editor
-    }
 
     /**
      * Get name of this plugin
@@ -50,7 +34,6 @@ export default class ContentModelEditPlugin
     initialize(editor: IEditor) {
         // TODO: Later we may need a different interface for Content Model editor plugin
         this.editor = editor as IContentModelEditor;
-        this.editor.getDocument().addEventListener('selectionchange', this.onNativeSelectionChange);
     }
 
     /**
@@ -59,19 +42,7 @@ export default class ContentModelEditPlugin
      * called, plugin should not call to any editor method since it will result in error.
      */
     dispose() {
-        if (this.editor) {
-            this.editor
-                .getDocument()
-                .removeEventListener('selectionchange', this.onNativeSelectionChange);
-            this.editor = null;
-        }
-    }
-
-    /**
-     * Get plugin state object
-     */
-    getState(): ContentModelEditPluginState {
-        return this.state;
+        this.editor = null;
     }
 
     /**
@@ -86,15 +57,6 @@ export default class ContentModelEditPlugin
                 case PluginEventType.KeyDown:
                     this.handleKeyDownEvent(this.editor, event);
                     break;
-
-                case PluginEventType.Input:
-                case PluginEventType.SelectionChanged:
-                    this.reconcileSelection(this.editor);
-                    break;
-
-                case PluginEventType.ContentChanged:
-                    this.clearCachedModel(this.editor);
-                    break;
             }
         }
     }
@@ -105,7 +67,7 @@ export default class ContentModelEditPlugin
 
         if (rawEvent.defaultPrevented || event.handledByEditFeature) {
             // Other plugins already handled this event, so it is most likely content is already changed, we need to clear cached content model
-            this.clearCachedModel(editor);
+            editor.invalidateCache();
         } else {
             // TODO: Consider use ContentEditFeature and need to hide other conflict features that are not based on Content Model
             switch (which) {
@@ -115,37 +77,7 @@ export default class ContentModelEditPlugin
                     // No need to clear cache here since if we rely on browser's behavior, there will be Input event and its handler will reconcile cache
                     keyboardDelete(editor, rawEvent);
                     break;
-
-                case Keys.ENTER:
-                    // ENTER key will create new paragraph, so need to update cache to reflect this change
-                    this.clearCachedModel(editor);
-                    break;
-
-                default:
-                    if (isCharacterValue(rawEvent) || rawEvent.key == ProcessKey) {
-                        applyDefaultFormat(editor);
-                    }
-                    break;
             }
-        }
-    }
-
-    private onNativeSelectionChange = () => {
-        if (this.editor?.hasFocus()) {
-            this.reconcileSelection(this.editor);
-        }
-    };
-
-    private reconcileSelection(editor: IContentModelEditor, newRangeEx?: SelectionRangeEx) {
-        // TODO: Really do reconcile selection
-        this.state.cachedModel = undefined;
-        this.state.cachedRangeEx = undefined;
-    }
-
-    private clearCachedModel(editor: IContentModelEditor) {
-        if (!editor.isInShadowEdit()) {
-            this.state.cachedModel = undefined;
-            this.state.cachedRangeEx = undefined;
         }
     }
 }
@@ -154,8 +86,7 @@ export default class ContentModelEditPlugin
  * @internal
  * Create a new instance of ContentModelEditPlugin class.
  * This is mostly for unit test
- * @param state State of this plugin
  */
-export function createContentModelEditPlugin(state: ContentModelEditPluginState) {
-    return new ContentModelEditPlugin(state);
+export function createContentModelEditPlugin() {
+    return new ContentModelEditPlugin();
 }
