+++ packages-content-model/roosterjs-content-model-editor/lib/publicTypes/IContentModelEditor.ts
@@ -37,6 +37,11 @@ export interface IContentModelEditor extends IEditor {
     ): void;
 
     /**
+     * Notify editor the current cache may be invalid
+     */
+    invalidateCache(): void;
+
+    /**
      * Get default format as ContentModelSegmentFormat.
      * This is a replacement of IEditor.getDefaultFormat for Content Model.
      * @returns The default format
