+++ packages-content-model/roosterjs-content-model-editor/test/editor/plugins/ContentModelEditPluginTest.ts
@@ -1,42 +1,27 @@
-import * as formatWithContentModel from '../../../lib/publicApi/utils/formatWithContentModel';
 import * as keyboardDelete from '../../../lib/publicApi/editing/keyboardDelete';
-import * as pendingFormat from '../../../lib/modelApi/format/pendingFormat';
 import ContentModelEditPlugin from '../../../lib/editor/plugins/ContentModelEditPlugin';
-import { ContentModelEditPluginState } from '../../../lib/publicTypes/pluginState/ContentModelEditPluginState';
+import { EntityOperation, Keys, PluginEventType } from 'roosterjs-editor-types';
 import { IContentModelEditor } from '../../../lib/publicTypes/IContentModelEditor';
-import { Position } from 'roosterjs-editor-dom';
-import {
-    EntityOperation,
-    Keys,
-    PluginEventType,
-    SelectionRangeTypes,
-} from 'roosterjs-editor-types';
 
 describe('ContentModelEditPlugin', () => {
     let editor: IContentModelEditor;
-    let pluginState: ContentModelEditPluginState;
+    let invalidateCache: jasmine.Spy;
     let getContentModelDefaultFormat: jasmine.Spy;
-    let plugin: ContentModelEditPlugin;
 
     beforeEach(() => {
-        pluginState = {};
+        invalidateCache = jasmine.createSpy('invalidateCache');
         getContentModelDefaultFormat = jasmine
             .createSpy('getContentModelDefaultFormat')
             .and.returnValue({});
 
         editor = ({
+            invalidateCache,
             getContentModelDefaultFormat,
-            getDocument: () => document,
-            isInShadowEdit: () => false,
             getSelectionRangeEx: () =>
                 ({
                     type: -1,
                 } as any), // Force return invalid range to go through content model code
         } as any) as IContentModelEditor;
-    });
-
-    afterEach(() => {
-        plugin?.dispose();
     });
 
     describe('onPluginEvent', () => {
@@ -47,7 +32,7 @@ describe('ContentModelEditPlugin', () => {
         });
 
         it('Backspace', () => {
-            plugin = new ContentModelEditPlugin(pluginState);
+            const plugin = new ContentModelEditPlugin();
             const rawEvent = { which: Keys.BACKSPACE } as any;
 
             plugin.initialize(editor);
@@ -58,11 +43,11 @@ describe('ContentModelEditPlugin', () => {
             });
 
             expect(keyboardDeleteSpy).toHaveBeenCalledWith(editor, rawEvent);
-            expect(pluginState).toEqual({});
+            expect(invalidateCache).not.toHaveBeenCalled();
         });
 
         it('Delete', () => {
-            plugin = new ContentModelEditPlugin(pluginState);
+            const plugin = new ContentModelEditPlugin();
             const rawEvent = { which: Keys.DELETE } as any;
 
             plugin.initialize(editor);
@@ -73,11 +58,11 @@ describe('ContentModelEditPlugin', () => {
             });
 
             expect(keyboardDeleteSpy).toHaveBeenCalledWith(editor, rawEvent);
-            expect(pluginState).toEqual({});
+            expect(invalidateCache).not.toHaveBeenCalled();
         });
 
         it('Other key', () => {
-            plugin = new ContentModelEditPlugin(pluginState);
+            const plugin = new ContentModelEditPlugin();
             const rawEvent = { which: 41 } as any;
 
             plugin.initialize(editor);
@@ -88,11 +73,11 @@ describe('ContentModelEditPlugin', () => {
             });
 
             expect(keyboardDeleteSpy).not.toHaveBeenCalled();
-            expect(pluginState).toEqual({});
+            expect(invalidateCache).not.toHaveBeenCalled();
         });
 
         it('Default prevented', () => {
-            plugin = new ContentModelEditPlugin(pluginState);
+            const plugin = new ContentModelEditPlugin();
             const rawEvent = { which: Keys.DELETE, defaultPrevented: true } as any;
 
             plugin.initialize(editor);
@@ -102,11 +87,11 @@ describe('ContentModelEditPlugin', () => {
             });
 
             expect(keyboardDeleteSpy).not.toHaveBeenCalled();
-            expect(pluginState).toEqual({ cachedModel: undefined, cachedRangeEx: undefined });
+            expect(invalidateCache).toHaveBeenCalled();
         });
 
         it('Trigger entity event first', () => {
-            plugin = new ContentModelEditPlugin(pluginState);
+            const plugin = new ContentModelEditPlugin();
             const wrapper = 'WRAPPER' as any;
 
             plugin.initialize(editor);
@@ -138,11 +123,11 @@ describe('ContentModelEditPlugin', () => {
             expect(keyboardDeleteSpy).toHaveBeenCalledWith(editor, {
                 which: Keys.DELETE,
             } as any);
-            expect(pluginState).toEqual({});
+            expect(invalidateCache).not.toHaveBeenCalled();
         });
 
         it('SelectionChanged event should clear cached model', () => {
-            plugin = new ContentModelEditPlugin(pluginState);
+            const plugin = new ContentModelEditPlugin();
 
             plugin.initialize(editor);
             plugin.onPluginEvent({
@@ -150,11 +135,11 @@ describe('ContentModelEditPlugin', () => {
                 selectionRangeEx: null!,
             });
 
-            expect(pluginState).toEqual({ cachedModel: undefined, cachedRangeEx: undefined });
+            expect(invalidateCache).not.toHaveBeenCalled();
         });
 
         it('keyboardDelete returns false', () => {
-            plugin = new ContentModelEditPlugin(pluginState);
+            const plugin = new ContentModelEditPlugin();
 
             keyboardDeleteSpy.and.returnValue(false);
 
@@ -164,354 +149,7 @@ describe('ContentModelEditPlugin', () => {
                 selectionRangeEx: null!,
             });
 
-            expect(pluginState).toEqual({
-                cachedModel: undefined,
-                cachedRangeEx: undefined,
-            });
+            expect(invalidateCache).not.toHaveBeenCalled();
         });
-    });
-});
-
-describe('ContentModelEditPlugin for default format', () => {
-    let editor: IContentModelEditor;
-    let contentDiv: HTMLDivElement;
-    let getSelectionRangeEx: jasmine.Spy;
-    let getPendingFormatSpy: jasmine.Spy;
-    let setPendingFormatSpy: jasmine.Spy;
-    let cacheContentModelSpy: jasmine.Spy;
-    let addUndoSnapshotSpy: jasmine.Spy;
-    let plugin: ContentModelEditPlugin;
-    let pluginState: ContentModelEditPluginState;
-
-    beforeEach(() => {
-        pluginState = {};
-        setPendingFormatSpy = spyOn(pendingFormat, 'setPendingFormat');
-        getPendingFormatSpy = spyOn(pendingFormat, 'getPendingFormat');
-        getSelectionRangeEx = jasmine.createSpy('getSelectionRangeEx');
-        cacheContentModelSpy = jasmine.createSpy('cacheContentModel');
-        addUndoSnapshotSpy = jasmine.createSpy('addUndoSnapshot');
-
-        contentDiv = document.createElement('div');
-
-        editor = ({
-            contains: (e: Node) => contentDiv != e && contentDiv.contains(e),
-            getSelectionRangeEx,
-            getContentModelDefaultFormat: () => ({
-                fontFamily: 'Arial',
-            }),
-            cacheContentModel: cacheContentModelSpy,
-            addUndoSnapshot: addUndoSnapshotSpy,
-            getDocument: () => document,
-        } as any) as IContentModelEditor;
-    });
-
-    afterEach(() => {
-        plugin?.dispose();
-    });
-
-    it('Collapsed range, text input, under editor directly', () => {
-        plugin = new ContentModelEditPlugin(pluginState);
-        const rawEvent = { key: 'a' } as any;
-
-        getSelectionRangeEx.and.returnValue({
-            type: SelectionRangeTypes.Normal,
-            ranges: [
-                {
-                    collapsed: true,
-                    startContainer: contentDiv,
-                    startOffset: 0,
-                },
-            ],
-        });
-
-        spyOn(formatWithContentModel, 'formatWithContentModel').and.callFake(
-            (_1: any, _2: any, callback: Function) => {
-                callback({
-                    blockGroupType: 'Document',
-                    blocks: [
-                        {
-                            blockType: 'Paragraph',
-                            format: {},
-                            isImplicit: true,
-                            segments: [
-                                {
-                                    segmentType: 'SelectionMarker',
-                                    format: {},
-                                    isSelected: true,
-                                },
-                            ],
-                        },
-                    ],
-                });
-            }
-        );
-
-        plugin.initialize(editor);
-
-        plugin.onPluginEvent({
-            eventType: PluginEventType.KeyDown,
-            rawEvent,
-        });
-
-        expect(setPendingFormatSpy).toHaveBeenCalledWith(
-            editor,
-            { fontFamily: 'Arial' },
-            new Position(contentDiv, 0)
-        );
-    });
-
-    it('Expanded range, text input, under editor directly', () => {
-        plugin = new ContentModelEditPlugin(pluginState);
-        const rawEvent = { key: 'a' } as any;
-
-        getSelectionRangeEx.and.returnValue({
-            type: SelectionRangeTypes.Normal,
-            ranges: [
-                {
-                    collapsed: false,
-                    startContainer: contentDiv,
-                    startOffset: 0,
-                },
-            ],
-        });
-
-        spyOn(formatWithContentModel, 'formatWithContentModel').and.callFake(
-            (_1: any, _2: any, callback: Function) => {
-                callback({
-                    blockGroupType: 'Document',
-                    blocks: [
-                        {
-                            blockType: 'Paragraph',
-                            format: {},
-                            isImplicit: true,
-                            segments: [
-                                {
-                                    segmentType: 'Text',
-                                    format: {},
-                                    text: 'test',
-                                    isSelected: true,
-                                },
-                            ],
-                        },
-                    ],
-                });
-            }
-        );
-
-        plugin.initialize(editor);
-
-        plugin.onPluginEvent({
-            eventType: PluginEventType.KeyDown,
-            rawEvent,
-        });
-
-        expect(setPendingFormatSpy).not.toHaveBeenCalled();
-        expect(addUndoSnapshotSpy).toHaveBeenCalledTimes(1);
-    });
-
-    it('Collapsed range, IME input, under editor directly', () => {
-        plugin = new ContentModelEditPlugin(pluginState);
-        const rawEvent = { key: 'Process' } as any;
-
-        getSelectionRangeEx.and.returnValue({
-            type: SelectionRangeTypes.Normal,
-            ranges: [
-                {
-                    collapsed: true,
-                    startContainer: contentDiv,
-                    startOffset: 0,
-                },
-            ],
-        });
-
-        spyOn(formatWithContentModel, 'formatWithContentModel').and.callFake(
-            (_1: any, _2: any, callback: Function) => {
-                callback({
-                    blockGroupType: 'Document',
-                    blocks: [
-                        {
-                            blockType: 'Paragraph',
-                            format: {},
-                            isImplicit: true,
-                            segments: [
-                                {
-                                    segmentType: 'SelectionMarker',
-                                    format: {},
-                                    isSelected: true,
-                                },
-                            ],
-                        },
-                    ],
-                });
-            }
-        );
-
-        plugin.initialize(editor);
-
-        plugin.onPluginEvent({
-            eventType: PluginEventType.KeyDown,
-            rawEvent,
-        });
-
-        expect(setPendingFormatSpy).toHaveBeenCalledWith(
-            editor,
-            { fontFamily: 'Arial' },
-            new Position(contentDiv, 0)
-        );
-    });
-
-    it('Collapsed range, other input, under editor directly', () => {
-        plugin = new ContentModelEditPlugin(pluginState);
-        const rawEvent = { key: 'Up' } as any;
-
-        getSelectionRangeEx.and.returnValue({
-            type: SelectionRangeTypes.Normal,
-            ranges: [
-                {
-                    collapsed: true,
-                    startContainer: contentDiv,
-                    startOffset: 0,
-                },
-            ],
-        });
-
-        spyOn(formatWithContentModel, 'formatWithContentModel').and.callFake(
-            (_1: any, _2: any, callback: Function) => {
-                callback({
-                    blockGroupType: 'Document',
-                    blocks: [
-                        {
-                            blockType: 'Paragraph',
-                            format: {},
-                            isImplicit: true,
-                            segments: [
-                                {
-                                    segmentType: 'SelectionMarker',
-                                    format: {},
-                                    isSelected: true,
-                                },
-                            ],
-                        },
-                    ],
-                });
-            }
-        );
-
-        plugin.initialize(editor);
-
-        plugin.onPluginEvent({
-            eventType: PluginEventType.KeyDown,
-            rawEvent,
-        });
-
-        expect(setPendingFormatSpy).not.toHaveBeenCalled();
-    });
-
-    it('Collapsed range, normal input, not under editor directly, no style', () => {
-        plugin = new ContentModelEditPlugin(pluginState);
-        const rawEvent = { key: 'a' } as any;
-        const div = document.createElement('div');
-
-        contentDiv.appendChild(div);
-
-        getSelectionRangeEx.and.returnValue({
-            type: SelectionRangeTypes.Normal,
-            ranges: [
-                {
-                    collapsed: true,
-                    startContainer: div,
-                    startOffset: 0,
-                },
-            ],
-        });
-
-        spyOn(formatWithContentModel, 'formatWithContentModel').and.callFake(
-            (_1: any, _2: any, callback: Function) => {
-                callback({
-                    blockGroupType: 'Document',
-                    blocks: [
-                        {
-                            blockType: 'Paragraph',
-                            format: {},
-                            segments: [
-                                {
-                                    segmentType: 'SelectionMarker',
-                                    format: {},
-                                    isSelected: true,
-                                },
-                            ],
-                        },
-                    ],
-                });
-            }
-        );
-
-        plugin.initialize(editor);
-
-        plugin.onPluginEvent({
-            eventType: PluginEventType.KeyDown,
-            rawEvent,
-        });
-
-        expect(setPendingFormatSpy).toHaveBeenCalledWith(
-            editor,
-            { fontFamily: 'Arial' },
-            new Position(div, 0)
-        );
-    });
-
-    it('Collapsed range, text input, under editor directly, has pending format', () => {
-        plugin = new ContentModelEditPlugin(pluginState);
-        const rawEvent = { key: 'a' } as any;
-
-        getSelectionRangeEx.and.returnValue({
-            type: SelectionRangeTypes.Normal,
-            ranges: [
-                {
-                    collapsed: true,
-                    startContainer: contentDiv,
-                    startOffset: 0,
-                },
-            ],
-        });
-
-        spyOn(formatWithContentModel, 'formatWithContentModel').and.callFake(
-            (_1: any, _2: any, callback: Function) => {
-                callback({
-                    blockGroupType: 'Document',
-                    blocks: [
-                        {
-                            blockType: 'Paragraph',
-                            format: {},
-                            isImplicit: true,
-                            segments: [
-                                {
-                                    segmentType: 'SelectionMarker',
-                                    format: {},
-                                    isSelected: true,
-                                },
-                            ],
-                        },
-                    ],
-                });
-            }
-        );
-
-        getPendingFormatSpy.and.returnValue({
-            fontSize: '10pt',
-        });
-
-        plugin.initialize(editor);
-
-        plugin.onPluginEvent({
-            eventType: PluginEventType.KeyDown,
-            rawEvent,
-        });
-
-        expect(setPendingFormatSpy).toHaveBeenCalledWith(
-            editor,
-            { fontFamily: 'Arial', fontSize: '10pt' },
-            new Position(contentDiv, 0)
-        );
     });
 });
